+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 ## Unreleased
 
+### Added
+
+- Multiple exports are now allowed if the build daemon is detected to have
+  version 0.13 of Buildkit or newer. (https://github.com/pulumi/pulumi-docker-build/issues/21)
+
 ### Changed
 
 - Upgraded buildx from 0.16.0 to 0.18.0.
 
 ### Fixed
+
 - Custom `# syntax=` directives no longer cause validation errors. (https://github.com/pulumi/pulumi-docker-build/issues/300)
 
 ## 0.0.7 (2024-10-16)